Move GRUB to pyanaconda.bootloader.grub
Move the GRUB class and the related code to a new file.
